refactor: pass on current focus
